requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.8
django-filter==2.3.0
django-import-export==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.4
packaging==20.4
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
